Move xxx.String() implementations in dbg_helper.go
